fix: quiz 답변 컴포넌트 분리 및 result api 엔드포인트 수정
